fix: close popups properly, move types to shared folder, update imports
